Root Shell UCE
Adds a root shell UCE with a modified build script to escalate
privilege. Includes the `root-me` binary (and source) which allows
running a command with args as root (when called from the UCE),
as well as socat and some other utilities for poking around. The
built UCE doesn't need the ext4 storage padded onto it, but that
can be changed in a later revision.

Adds `un-uce` for extracting a UCE. It's just a thin wrapper around
`unsquashfs`.